controllerのdomainを用いての書き換え
