--- cmd/controller/models/root.go
@@ -1,7 +1,0 @@
-package models
-
-// Root yamlのルート
-type Root struct {
-	Config *Config  `yaml:"config"`
-	Tables []*Table `yaml:"tables"`
-}
